Based on the infinity_demo.py, separate controller design from meshcat simulation.
