+++ drake/bouncing_ball/simulation_connector.py
@@ -1,8 +1,21 @@
-## simulation_connector.py
-#   Description:
-#       This function is meant to provide a connection module to the meshcat simulator for our Kinova arm.
-#       This is built based on infinity_simulation.py
-##
+"""_summary_
+    This function is meant to provide a connection module to the meshcat simulator for our Kinova arm.
+    This is built based on infinity_simulation.py
+    It is worth considering that in "infinity_simulation.py", there are lines of codes being commented out.
+    They are extensional functions. In this script, only necessary functions will be remained for the
+    purposes of clarity and minimizing.
+
+    In bouncing ball controlling, the end effector is replaced by a paddle. No actuator or sensor maneuvers the paddle.
+    Therefore, any function concerning with gripper controllers is deleted.
+
+Raises:
+    RuntimeError: _description_
+    ConnectionError: The script failed to connect to meshcat
+
+Returns:
+    _type_: _description_
+    image:
+"""
 
 ###########
 # Imports #
@@ -18,189 +31,19 @@ from meshcat.servers.zmqserver import start_zmq_server_as_subprocess
 proc, zmq_url, web_url = start_zmq_server_as_subprocess(server_args=server_args)
 
 
-## import kinova drake which is not in the subdirectories here
-from kinova_drake.kinova_station import KinovaStationHardwareInterface, EndEffectorTarget, GripperTarget, KinovaStation
-#from kinova_drake.controllers import CommandSequenceController, CommandSequence, Command
-from kinova_drake.observers.camera_viewer import CameraViewer
+# Import kinova drake which is not in the subdirectories here
+import sys
+sys.path.append('/root/kinova_drake/')
+from kinova_station import KinovaStationHardwareInterface, EndEffectorTarget, GripperTarget, KinovaStation
+from observers.camera_viewer import CameraViewer
 
-####################
-# Control commands #
-####################
-
-def setup_gripper_command_systems(gripper_command_type,builder,station):
-    # Set gripper command
-    if gripper_command_type == GripperTarget.kPosition:
-        q_grip_des = np.array([0.0])   # open at 0, closed at 1
-        gripper_target_source = builder.AddSystem(ConstantVectorSource(q_grip_des))
-
-    elif gripper_command_type == GripperTarget.kVelocity:
-        v_grip_des = np.array([1.0])
-        gripper_target_source = builder.AddSystem(ConstantVectorSource(v_grip_des))
-
-    # Send gripper command and type
-    gripper_target_type_source = builder.AddSystem(ConstantValueSource(
-                                            AbstractValue.Make(gripper_command_type)))
-    builder.Connect(
-            gripper_target_type_source.get_output_port(),
-            station.GetInputPort("gripper_target_type"))
-
-    builder.Connect(
-            gripper_target_source.get_output_port(),
-            station.GetInputPort("gripper_target"))
-
-    gripper_target_source.set_name("gripper_command_source")
-    gripper_target_type_source.set_name("gripper_type_source")
-
-def create_end_effector_target(ee_command_type,builder,station):
-    """
-    create_end_effector_target
-    Description:
-        Creates the type of commands that are sent to the arm and the gripper.
-        Options are kPose, kTwist, or kWrench.
-    """
-    #ee_command_type = EndEffectorTarget.kPose      # kPose, kTwist, or kWrench
-
-    # Set (constant) command to send to the system
-    if ee_command_type == EndEffectorTarget.kPose:
-        pose_des = np.array([np.pi,0.0,0.0,
-                                0.6,0.0,0.2])
-        target_source = builder.AddSystem(ConstantVectorSource(pose_des))
-    elif ee_command_type == EndEffectorTarget.kTwist:
-        twist_des = np.array([0.0,0.1,0.0,
-                            0.0,0.0,0.0])
-        target_source = builder.AddSystem(ConstantVectorSource(twist_des))
-    elif ee_command_type == EndEffectorTarget.kWrench:
-        wrench_des = np.array([0.0,0.0,0.0,
-                                0.0,0.0,0.0])
-        target_source = builder.AddSystem(ConstantVectorSource(wrench_des))
-    else:
-        raise RuntimeError("invalid end-effector target type")
-
-    # Create the system which outputs the End Effector Command's Type
-    target_type_source = builder.AddSystem(ConstantValueSource(AbstractValue.Make(ee_command_type)))
-
-    # Name the new systems
-    target_source.set_name("ee_command_source")
-    target_type_source.set_name("ee_type_source")
-
-    # Connecting the New Systems (target_type_source and target_source)
-    builder.Connect(
-            target_type_source.get_output_port(),
-            station.GetInputPort("ee_target_type")) # connects to a port ON THE STATION with the given name
-    builder.Connect(
-            target_source.get_output_port(),
-            station.GetInputPort("ee_target"))
-
-def setup_infinity_command_sequence():
-    """
-    Description:
-        Creates the command sequence that we need to achieve the infinity sequence.
-    """
-    # Constants
-    num_points_in_discretized_curve = 7
-    infinity_center = np.array([0.6, 0.0, 0.5])
-    infinity_center_pose = np.zeros((6,))
-    infinity_center_pose[:3] = np.array([np.pi/2,0.0,0.5*np.pi])
-    infinity_center_pose[3:] = infinity_center
-
-    curve_radius = 0.2
-    curve_duration = 6.0
-
-    infinity_left_lobe_center = infinity_center - np.array([0.0,0.25,0.0])
-    infinity_left_lobe_center_pose = np.zeros((6,))
-    infinity_left_lobe_center_pose[:3] = np.array([np.pi/2,0.0,0.5*np.pi])
-    infinity_left_lobe_center_pose[3:] = infinity_left_lobe_center 
-
-    infinity_right_lobe_center = infinity_center + np.array([0.0,0.25,0.0])
-    infinity_right_lobe_center_pose = np.zeros((6,))
-    infinity_right_lobe_center_pose[:3] = np.array([np.pi/2,0.0,0.5*np.pi])
-    infinity_right_lobe_center_pose[3:] = infinity_right_lobe_center
-
-    # Create the command sequence object
-    cs = CommandSequence([])
-
-    # 1. Initial Command
-    cs.append(Command(
-        name="centering",
-        target_pose=infinity_center_pose,
-        duration=6,
-        gripper_closed=False))    
-
-    # 2. Lower Left Part of Infinity
-    cs.append(Command(
-        name="lower_left",
-        target_pose=infinity_left_lobe_center_pose - np.array([0,0,0,0,0,curve_radius]),
-        duration=4,
-        gripper_closed=False
-    ))
-    # 3. Execute curve for left lobe
-    theta_list = np.linspace(np.pi/2,np.pi*1.5,num=num_points_in_discretized_curve)
-    flipped_theta_list = np.flipud(theta_list)
-    for theta_index in range(1,len(flipped_theta_list)):
-        theta = flipped_theta_list[theta_index]
-        cs.append(Command(
-            name="left_curve"+str(theta_index),
-            target_pose=infinity_left_lobe_center_pose + np.array([0.0,0.0,0.0,0,curve_radius*np.cos(theta),curve_radius*np.sin(theta)]),
-            duration= curve_duration/(num_points_in_discretized_curve),
-            gripper_closed = False
-        ))
-
-    # 4. Go to lower right Part of infinity (through center)
-    cs.append(Command(
-        name="left_curve_to_right",
-        target_pose = infinity_right_lobe_center_pose - np.array([0,0,0,0,0,curve_radius]),
-        duration=8,
-        gripper_closed=False
-    ))
-    # 5. Execute curve for right lobe
-    theta_list = np.linspace(np.pi*1.5,np.pi*2.5,num=num_points_in_discretized_curve)
-    for theta_index in range(1,len(theta_list)):
-        theta = theta_list[theta_index]
-        cs.append(Command(
-            name="right_curve"+str(theta_index),
-            target_pose=infinity_right_lobe_center_pose + np.array([0.0,0.0,0.0,0,curve_radius*np.cos(theta),curve_radius*np.sin(theta)]),
-            duration= curve_duration/(num_points_in_discretized_curve),
-            gripper_closed = False
-        ))
-
-    #6. Get Back to Center
-    cs.append(Command(
-        name="right_curve_to_home",
-        target_pose = infinity_center_pose,
-        duration=4,
-        gripper_closed=False
-    ))
-
-    return cs
-
-def setup_controller_and_connect_to_station(cs,builder,station):
-    """
-    Description:
-        Defines the controller (PID) which is a CommandSequenceController as defined in
-        kinova_drake.
-    Inputs:
-        cs = A CommandSequence object which helps define the CommandSequenceController.
-    """
-
-    # Create the controller and connect inputs and outputs appropriately
-    #Kp = 10*np.eye(6)
-    Kp = np.diag([10,10,10,2,2,2])
-    Kd = 2*np.sqrt(Kp)
-
-    controller = builder.AddSystem(CommandSequenceController(
-        cs,
-        command_type=EndEffectorTarget.kTwist,  # Twist commands seem most reliable in simulation
-        Kp=Kp,
-        Kd=Kd))
-    controller.set_name("controller")
-    controller.ConnectToStation(builder, station)
-
-    return controller
-
+# Import the controller you designed
+import control_module
 
 ####################
 # Helper Functions #
 ####################
+
 def add_loggers_to_system(builder,station):
     # Loggers force certain outputs to be computed
     wrench_logger = LogVectorOutput(station.GetOutputPort("measured_ee_wrench"),builder)
@@ -224,23 +67,23 @@ def create_scenario():
     Usage:
     """
 
-    # Constants
+    # Determine the type of the end effector (gripper)
     gripper_type = '2f_85'
 
+    ## Build the environment
     # Start with the Kinova Station object
     station = KinovaStation(time_step=0.001,n_dof=6)
-
-    # station.AddArm()
-    # station.Add2f85Gripper()
     station.AddArmWith2f85Gripper()
-
     station.AddGround()
     station.AddCamera()
-    # station.SetupSinglePegScenario(gripper_type=gripper_type, arm_damping=False)
+    #station.SetupSinglePegScenario(gripper_type=gripper_type, arm_damping=False)
+
+    # Create environment in MeshCat
     station.ConnectToMeshcatVisualizer(zmq_url="tcp://127.0.0.1:6000")
 
     station.Finalize() # finalize station (a diagram in and of itself)
 
+    ## Build the Kinova
     # Start assembling the overall system diagram
     builder = DiagramBuilder()
     station = builder.AddSystem(station)
@@ -253,8 +96,7 @@ def create_scenario():
     add_loggers_to_system(builder,station)
 
     # Setup Controller
-    cs = setup_infinity_command_sequence()
-    controller = setup_controller_and_connect_to_station(cs,builder,station)
+    controller = control_module.setup_example_controller_and_connect_to_station(builder,station)
 
     # Build the system diagram
     diagram = builder.Build()
@@ -277,11 +119,12 @@ show_station_diagram = False
 
 # Make a plot of the diagram for this example, where only the inputs
 # and outputs of the station are shown
-show_toplevel_diagram = False
+show_toplevel_diagram = True
 
 # Which gripper to use (hande or 2f_85)
 gripper_type = "2f_85"
 
+# Option of simulation
 run = True
 
 ##########################
@@ -309,9 +152,12 @@ if run:
     time_step = 0.025
     #ResetIntegratorFromFlags(simulator, integration_scheme, time_step)
 
-    # Run simulation
-    simulator.Initialize()
-    simulator.AdvanceTo(40.0)
+    try:
+        # Run simulation
+        simulator.Initialize()
+        simulator.AdvanceTo(40.0)
+    except Exception as e:
+        print(str(e))
 
 while True:
     1
